Refactor docpage, sidebar, navbar (#409)
