+++ pcweb/components/sidebar/sidebar.py
@@ -14,17 +14,6 @@ from .sidebar_items.learn import learn, frontend, backend, hosting
 from .sidebar_items.component_lib import get_component_link, component_lib, other_libs
 from .sidebar_items.reference import api_reference, recipes, tutorials
 
-def sidebar_section(name):
-    return rx.heading(
-        rx.span("[ ", color="#DACEEE"),
-        name,
-        rx.span(" ]", color="#DACEEE"),
-        style=heading_style3,
-        margin_top="1em",
-        margin_left="0em",
-        padding_left="1em",
-    )
-
 
 def sidebar_link(*children, **props):
     """Create a sidebar link that closes the sidebar when clicked."""
@@ -207,6 +196,61 @@ def get_prev_next(url):
     return None, None
 
 
+def sidebar_category(name, color, index):
+    return rx.box(
+            rx.text(
+                name,
+                font_weight= fw["section"],
+                font_size= styles.TEXT_FONT_SIZE,
+                color= "#696287",  
+            ),    
+            on_click=lambda: SidebarState.set_sidebar_index(index),
+            background=rx.cond(
+                SidebarState.sidebar_index == index,
+                "#F5EFFE",
+                "transparent",
+            ),
+            align_items="left",
+            padding_y="0.5em",
+            padding_x="0.5em",
+            border_radius="0.5em",
+            border="2px solid #F4F3F6",
+            width="100%",
+        )
+
+def sidebar_section(name):
+    return rx.heading(
+        rx.span("[ ", color="#DACEEE"),
+        name,
+        rx.span(" ]", color="#DACEEE"),
+        style=heading_style3,
+        margin_top="1em",
+        margin_left="0em",
+        padding_left="1em",
+    )
+
+def create_sidebar_section(section_title, items, index, url):
+    return rx.vstack(
+        sidebar_section(section_title),
+        rx.accordion(
+            *[
+                sidebar_item_comp(
+                    item=item,
+                    index=[-1],
+                    url=url,
+                )
+                for item in items
+            ],
+            allow_multiple=True,
+            default_index=index if index is not None else [],
+            width="100%",
+        ),
+        padding_x="0em",
+        width="100%",
+        align_items="left",
+    )
+
+
 @rx.memo
 def sidebar_comp(
     url: str,
@@ -216,141 +260,32 @@ def sidebar_comp(
     backend_index: list[int],
     hosting_index: list[int],
     other_libs_index: list[int],
+    api_reference_index: list[int],
+    recipes_index: list[int],
+    tutorials_index: list[int],
 ):
     return rx.vstack(
-        rx.box(
-            rx.text(
-                "Learn",
-                font_weight= fw["section"],
-                font_size= styles.TEXT_FONT_SIZE,
-                color= "#696287",  
-            ),    
-            on_click=lambda: SidebarState.set_sidebar_index(0),
-            background=rx.cond(
-                SidebarState.sidebar_index == 0,
-                "#F5EFFE",
-                "transparent",
-            ),
-            align_items="left",
-            padding_y="0.5em",
-            padding_x="0.5em",
-            border_radius="0.5em",
-            border="2px solid #F4F3F6",
-            width="100%",
-        ),
-        rx.box(
-            rx.text(
-                "Reference",
-                font_weight= fw["section"],
-                font_size= styles.TEXT_FONT_SIZE,
-                color= "#696287",  
-            ), 
-            on_click=lambda: SidebarState.set_sidebar_index(1),
-            background=rx.cond(
-                SidebarState.sidebar_index == 1,
-                "#F5EFFE",
-                "transparent",
-            ),
-            align_items="left",
-            padding_y="0.5em",
-            padding_x="0.5em",
-            border_radius="0.5em",
-            border="2px solid #F4F3F6",
-            width="100%",
-        ),
+        sidebar_category("Learn", "red", 0),
+        sidebar_category("Components", "red", 1),
+        sidebar_category("API Reference", "red", 2),
         rx.divider(),
-        rx.cond(
-            SidebarState.sidebar_index == 0,
-            rx.vstack(
-                sidebar_section("Onboarding"),
-                rx.accordion(
-                    *[
-                        sidebar_item_comp(
-                            item=item,
-                            index=[-1],
-                            url=url,
-                        )
-                        for item in learn
-                    ],
-                    allow_multiple=True,
-                    default_index=learn_index if learn_index is not None else [],
-                    width="100%",
-                ),
-                sidebar_section("UI"),
-                rx.accordion(
-                    *[
-                        sidebar_item_comp(
-                            item=item,
-                            index=[-1],
-                            url=url,
-                        )
-                        for item in frontend
-                    ],
-                    allow_multiple=True,
-                    default_index=frontend_index if frontend_index is not None else [],
-                    width="100%",
-                ),
-                sidebar_section("State"),
-                rx.accordion(
-                    *[
-                        sidebar_item_comp(
-                            item=item,
-                            index=[-1],
-                            url=url,
-                        )
-                        for item in backend
-                    ],
-                    allow_multiple=True,
-                    default_index=backend_index if backend_index is not None else [],
-                    width="100%",
-                ),
-                sidebar_section("Hosting"),
-                rx.accordion(
-                    *[
-                        sidebar_item_comp(
-                            item=item,
-                            index=[-1],
-                            url=url,
-                        )
-                        for item in hosting
-                    ],
-                    allow_multiple=True,
-                    default_index=hosting_index if hosting_index is not None else [],
-                    width="100%",
-                ),
-                padding_x="0em",
-                width="100%",
-                align_items="left",
-            ),
-            rx.vstack(
-                sidebar_section("Core"),
-                rx.accordion(
-                    *[
-                        sidebar_item_comp(item=item, url=url, index=component_lib_index)
-                        for item in component_lib
-                    ],
-                    allow_multiple=True,
-                    default_index=component_lib_index
-                    if component_lib_index is not None
-                    else [],
-                    width="100%",
-                ),
-                sidebar_section("Other Libraries"),
-                rx.accordion(
-                    *[
-                        sidebar_item_comp(item=item, url=url, index=other_libs_index)
-                        for item in other_libs
-                    ],
-                    allow_multiple=True,
-                    default_index=other_libs_index
-                    if other_libs_index is not None
-                    else [],
-                    width="100%",
-                ),
-                padding_x="0em",
-                width="100%",
-                align_items="left",
-            ),
+        rx.match(
+            SidebarState.sidebar_index,
+            (0, rx.vstack(
+                create_sidebar_section("Onboarding", learn, learn_index, url),
+                create_sidebar_section("UI", frontend, frontend_index, url),
+                create_sidebar_section("State", backend, backend_index, url),
+                create_sidebar_section("Hosting", hosting, hosting_index, url)
+            )),
+            (1, rx.vstack(
+                create_sidebar_section("Core Components", component_lib, component_lib_index, url),
+                create_sidebar_section("Other Libraries", other_libs, other_libs_index, url)         
+            )),
+            (2, rx.vstack(
+                create_sidebar_section("API Reference", api_reference, api_reference_index, url),
+                create_sidebar_section("Recipes", recipes, recipes_index, url),
+                create_sidebar_section("Tutorials", tutorials, tutorials_index, url)       
+            )),
         ),
         align_items="left",
         overflow_y="scroll",
@@ -379,6 +314,9 @@ def sidebar(url=None) -> rx.Component:
     backend_index = calculate_index(backend, url)
     hosting_index = calculate_index(hosting, url)
     other_libs_index = calculate_index(other_libs, url)
+    api_reference_index = calculate_index(api_reference, url)
+    recipes_index = calculate_index(recipes, url)
+    tutorials_index = calculate_index(tutorials, url)
 
     return rx.box(
         sidebar_comp(
@@ -389,6 +327,9 @@ def sidebar(url=None) -> rx.Component:
             backend_index=backend_index,
             hosting_index=hosting_index,
             other_libs_index=other_libs_index,
+            api_reference_index=api_reference_index,
+            recipes_index=recipes_index,
+            tutorials_index=tutorials_index,
         ),
         width="100%",
     )
