+++ pcweb/pcweb.py
@@ -19,7 +19,7 @@ app = rx.App(
     style=styles.BASE_STYLE,
     stylesheets=styles.STYLESHEETS,
     theme=rdxt.theme(
-        appearance="dark", has_background=True, radius="large", accent_color="teal"
+        appearance="light", has_background=True, radius="large", accent_color="teal"
     ),
     head_components=[
         rx.script(src="https://www.googletagmanager.com/gtag/js?id=G-4T7C8ZD9TR"),
