+++ pcweb/templates/docpage/__init__.py
@@ -1,0 +1,1 @@
+from .docpage import *
