+++ pcweb/templates/docpage/blocks/__init__.py
@@ -1,0 +1,4 @@
+from .headings import *
+from .demo import *
+from .code import *
+from .typography import *
