+++ pcweb/templates/docpage/blocks/typography.py
@@ -1,0 +1,64 @@
+"""Typography blocks for doc pages."""
+
+import reflex as rx
+from pcweb import styles
+from pcweb.styles import colors as c
+
+def definition(title: str, *children) -> rx.Component:
+    """Create a definition for a doc page.
+
+    Args:
+        title: The title of the definition.
+        children: The children to display.
+
+    Returns:
+        The styled definition.
+    """
+    return rx.box(
+        rx.heading(title, font_size="1em", margin_bottom="0.5em", font_weight="bold"),
+        *children,
+        padding="1em",
+        border=styles.DOC_BORDER,
+        border_radius=styles.DOC_BORDER_RADIUS,
+        _hover={
+            "box_shadow": styles.DOC_SHADOW_LIGHT,
+            "border": f"2px solid {c['violet'][200]}",
+        },
+    )
+
+@rx.memo
+def text_comp(text: rx.Var[str]) -> rx.Component:
+    return rx.text(text, margin_bottom="1em", font_size=styles.TEXT_FONT_SIZE)
+
+
+@rx.memo
+def code_comp(text: rx.Var[str]) -> rx.Component:
+    return rx.code(text, color="#1F1944", bg="#EAE4FD")
+
+
+def doclink(text: str, href: str, **props) -> rx.Component:
+    """Create a styled link for doc pages.
+
+    Args:
+        text: The text to display.
+        href: The link to go to.
+        props: Props to apply to the link.
+
+    Returns:
+        The styled link.
+    """
+    return rx.link(text, href=href, style=styles.LINK_STYLE, **props)
+
+
+def doclink2(text: str, **props) -> rx.Component:
+    """Create a styled link for doc pages.
+
+    Args:
+        text: The text to display.
+        href: The link to go to.
+        props: Props to apply to the link.
+
+    Returns:
+        The styled link.
+    """
+    return rx.link(text, style=styles.LINK_STYLE, **props)
